Moved scale function to a separate module
